+++ src/components/activities/ReviewCard.tsx
@@ -1,6 +1,7 @@
 import StarRatingDisplay from './StarRatingDisplay';
+import { format } from 'date-fns';
+import { ko } from 'date-fns/locale';
 
-// 임시 타입 작성 API 요청 시 변경될 수 있습니다.
 interface ReviewCardProps {
   nickname: string;
   rating: number;
@@ -10,13 +11,7 @@ interface ReviewCardProps {
 
 const ReviewCard = ({ nickname, rating, content, createdAt }: ReviewCardProps) => {
   const date = new Date(createdAt);
-  const viewDate = date
-    .toLocaleDateString('ko-KR', {
-      year: 'numeric',
-      month: 'numeric',
-      day: 'numeric',
-    })
-    .replace(/\.$/, '');
+  const viewDate = format(date, 'yyyy.M.d', { locale: ko });
   return (
     <div className='shadow-custom-5 flex h-auto w-auto flex-col gap-12 rounded-3xl bg-white p-20'>
       <div className='flex flex-col justify-center gap-4'>
