+++ src/components/activities/ReviewSection.tsx
@@ -6,6 +6,7 @@ import Pagination from '@/components/common/Pagination';
 import ReviewCard from '@/components/activities/ReviewCard';
 import type { ReviewResponse } from './Activities.types';
 import { fetchActivityReviews } from '@/lib/api/activities/index';
+import getSatisfactionText from '@/utils/activities/getSatisfactionText';
 
 interface ReviewSectionProps {
   activityId: number;
@@ -25,7 +26,7 @@ const ReviewSection = ({ activityId }: ReviewSectionProps) => {
         setLoading(true);
         const data = await fetchActivityReviews(activityId, currentPage, PAGE_SIZE);
         setReviewData(data);
-      } catch (err) {
+      } catch {
         setReviewData(null);
       } finally {
         setLoading(false);
@@ -36,17 +37,22 @@ const ReviewSection = ({ activityId }: ReviewSectionProps) => {
 
   // 임시 처리
   if (loading) return <p>로딩 중...</p>;
-  if (!reviewData) return <p> 리뷰를 불러오지 못하였습니다. </p>;
+  if (!reviewData)
+    return (
+      <section className='flex flex-col items-center justify-center gap-8'>
+        <h2 className='text-18-b self-start text-gray-950'>
+          체험 후기 <span className='text-[0.875rem] font-semibold text-[#79747E]'>0개</span>
+        </h2>
+        <p className='text-lg font-bold text-gray-950'>리뷰를 불러오지 못하였습니다.</p>
+        <div className='mb-10 flex items-center gap-0.5'>
+          <Image src='/icons/icon_star_on.svg' alt='별' width={16} height={16} />
+          <p className='text-14-body-m text-gray-700'>0개 후기</p>
+        </div>
+        <div className='w-full py-10 text-center text-gray-500'>리뷰를 불러오지 못하였습니다.</div>
+      </section>
+    );
 
   const totalPages = Math.ceil(reviewData.totalCount / PAGE_SIZE);
-
-  const getSatisfactionText = (rating: number) => {
-    if (rating >= 4.5) return '매우 만족';
-    if (rating >= 3.5) return '만족';
-    if (rating >= 2.5) return '보통';
-    if (rating >= 1.5) return '불만족';
-    return '매우 불만족';
-  };
 
   if (reviewData.totalCount === 0) {
     return (
